+++ spec/perf/perf_spec.rb
@@ -22,7 +22,7 @@ module DeadEnd
           107    constraints -> { Rails.application.config.non_production } do
           111    end
         ❯ 113    namespace :admin do
-        ❯ 116    match "/out_of_office(*path)", via: :all, to: redirect { |_params, req|
+        ❯ 116    match "/foobar(*path)", via: :all, to: redirect { |_params, req|
         ❯ 120    }
           121  end
       EOM